Moved files around to make Streakers go install-able.
